Fix pin initilization
Use cbi/sbi macros to initialize pins. This is more consistent, but also
before when set to active low the way I was initializing PORTB would
have all of PORTB except INPUT_PIN set to high. This works, but it's
kind of sloppy.

Also, note added to readme about the config header.